change toggle loop button to icon button
