+++ src/component/Controller/index.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
+import IconButton from "../Common/IconButton";
 import PlayButton from "./PlayButton";
+import LoopButton from "./LoopButton";
 import SeekBar from "./SeekBar";
-import ToggleLoop from "./ToggleLoop";
 
 import { formatTime } from "../../format";
-import IconButton from "../Common/IconButton";
 
 type Props = {
   duration: number;
@@ -38,7 +38,7 @@ const Controller: React.FC<Props> = ({
         <IconButton icon="skip_previous" onClick={playPrev} />
         <PlayButton isPaused={paused} play={play} pause={pause} />
         <IconButton icon="skip_next" onClick={playNext} />
-        <ToggleLoop loop={loop} setLoop={setLoop} />
+        <LoopButton loop={loop} setLoop={setLoop} />
         {formatTime(currentTime)}/{formatTime(duration)}
       </div>
       <SeekBar duration={duration} time={currentTime} seek={seek} />
